backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.4.6
openai==1.54.0
PyMuPDF==1.24.5
scikit-learn==1.5.0
pydantic==2.7.4
httpx==0.27.0
beautifulsoup4==4.12.3
playwright==1.44.0
youtube-search-python==1.6.6
language-tool-python==2.8
reportlab==4.0.7
